requirements.txt
redis
requests
python-telegram-bot
